Incorporate Scott's patch to eliminate _schema_py.py
